requirements.txt
Flask
google-generativeai==0.8.5
gunicorn
